refactor: 파일 경로를 절대경로로 수정
