+++ src/components/Post/index.ts
@@ -1,0 +1,2 @@
+export { default as Posts } from './Posts/Posts';
+export { default as PostItem } from './PostItem/PostItem';
