+++ src/pages/HiresPage.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react';
 import { useDispatch } from 'react-redux';
 import { fetchAllPost } from '@/store/posts';
-import Posts from '@/components/Posts';
+import { Posts } from '@/components';
 
 // Todo(홍중) : 임시 데이터, 팀 정보 관리하는곳에서 가져오기(12/13)
 const selectedTeam = { teamId: 3, grade: 'GENERAL' };
